+++ test/bdd-syntax/test.mjs
@@ -9,4 +9,8 @@ test(testDir.name, () => {
   expect(
     testDir.getFileContents('.features-gen/features/scenario-simple.feature.spec.js'),
   ).toContain('import { test } from "../../steps/fixtures.ts";');
+
+  expect(
+    testDir.isFileExists('.features-gen/features/background-no-scenarios.feature.spec.js'),
+  ).toEqual(false);
 });