+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## dev
+* fix spec generation for feature file with background without scenarios, [#104](https://github.com/vitalets/playwright-bdd/issues/104)
+
 ## 6.0.1
 * fix Cucumber reporters for non-default language